DMCDOT-602 Updating the header to follow the format described in an email from Scott.
"Hi Jack and Scott,

As a quick follow-up to the topic of the inspection report title from last Friday, I wanted to let you know that we will skip the cover page, no cover page.

As part of a header or as part of the footer, we'd like to see three additional pieces of information.

[inspection type] + Report
[structure key]
[inspection status]"